+++ custom_components/webserver_app/utils.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import os
-from typing import Any
 
 from homeassistant.core import HomeAssistant