+++ ShieldDNS/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.4.5 (2026-04-24)
 
 ### 📦 Dependencies
+
 - ⬆️ Update dependency FaserF/ShieldDNS to v1.7.0 [skip-tests] (#771) [skip-tests] ([`e124a4d6`](https://github.com/FaserF/hassio-addons/commit/e124a4d6ef139eb59216dabe4b730578ace82b87))
 - ⬆️ Update dependency coredns/coredns to v1.14.3 [skip-tests] (#768) [skip-tests] ([`c5194348`](https://github.com/FaserF/hassio-addons/commit/c519434854888ac489e0101e24f55dd785925047))
 - ⬆️ Update ghcr.io/hassio-addons/base Docker tag to v[20.1.0](https://github.com/hassio-addons/addon-base/releases/tag/20.1.0) [skip-tests] ([`709f7882`](https://github.com/FaserF/hassio-addons/commit/709f7882b67adc67ab6f64370f1d900e9a71b2e1))
@@ -13,6 +14,7 @@
 - ⬆️ Update ghcr.io/hassio-addons/base Docker tag to v[20.0.3](https://github.com/hassio-addons/addon-base/releases/tag/20.0.3) [skip-tests] ([`d765bff1`](https://github.com/FaserF/hassio-addons/commit/d765bff148faae0ef9dfe1a7a4634b4fdc598592))
 
 ### 🚀 Other
+
 - small improvement ([`6d9a62b0`](https://github.com/FaserF/hassio-addons/commit/6d9a62b0579c0d15dc0eab3e3eade0671ea0742d))
 - small improvement ([`9ac5fc69`](https://github.com/FaserF/hassio-addons/commit/9ac5fc69462c4811e416c8e5c518c79e172c1477))
 - small code improvements ([`485e807a`](https://github.com/FaserF/hassio-addons/commit/485e807a6b8e989216de479c89700d2889b1c030))
@@ -22,7 +24,6 @@
 - fix docker build ([`db25a412`](https://github.com/FaserF/hassio-addons/commit/db25a4124eab24641d43807fc7b97df694d6e758))
 - fix docker build in newer alpine version ([`7a040ec8`](https://github.com/FaserF/hassio-addons/commit/7a040ec8304f4f6ffc2a5dbbcc237ef19e9aca10))
 - fix docker build in newer alpine version ([`2fcb0e33`](https://github.com/FaserF/hassio-addons/commit/2fcb0e33989a51b0ce36a32544b0281b8d82279f))
-
 
 ## 2.4.3 (2026-04-11)